+++ src/types/transcript.ts
@@ -8,16 +8,46 @@ export type SpeakerInfo = {
 export type Speaker = {
   speaker: string;
   speakerInfo?: SpeakerInfo;
-  text?: string;
-  words?: string[];
-  startTimes?: number[];
+  text: string;
 };
 
-export type Transcript = {
+export type WordLevelSpeaker = {
+  speaker: string;
+  speakerInfo?: SpeakerInfo;
+  words: string[];
+  startTimes: number[];
+};
+
+/**
+ * A transcript of a single earnings call.
+ */
+export type WordLevelTimestampsTranscript = {
+  event: EarningsEvent;
+  speakers: WordLevelSpeaker[];
+};
+
+/**
+ * A transcript of a single earnings call.
+ */
+export type QuestionAndAnswersTranscript = {
+  event: EarningsEvent;
+  preparedRemarks: string;
+  questionsAndAnswers: string;
+};
+
+/**
+ * A basic transcript of a single earnings call.
+ */
+export type BasicTranscript = {
   event: EarningsEvent;
   text: string;
-  speakers?: Speaker[];
-  preparedRemarks?: string;
-  questionsAndAnswers?: string;
+};
+
+/**
+ * A transcript of a single earnings call with speaker groups.
+ */
+export type SpeakerGroups = {
+  event: EarningsEvent;
+  speakers: Speaker[];
   speakerNameMapV2?: { [key: string]: SpeakerInfo };
 };